+ fix JUnit-Tests
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@2128 5890f383-b09f-485f-8f59-22bc967d67fd

+++ UnitTests/CB_Core_Tests/src/CB_Core/CB_Core/DB/TestDB.java
@@ -31,10 +31,10 @@ public class TestDB extends Database
 		try
 		{
 			myDB.close();
+			myDB = null;
 		}
-		catch (SQLException e)
+		catch (Exception e)
 		{
-
 			e.printStackTrace();
 		}
 	}
@@ -67,7 +67,7 @@ public class TestDB extends Database
 
 		try
 		{
-			Connection myDB = DriverManager.getConnection("jdbc:sqlite:" + databasePath);
+			myDB = DriverManager.getConnection("jdbc:sqlite:" + databasePath);
 			myDB.commit();
 			myDB.close();
 
@@ -81,15 +81,11 @@ public class TestDB extends Database
 	@Override
 	public CoreCursor rawQuery(String sql, String[] args)
 	{
-
+		if (myDB == null) return null;
 		ResultSet rs = null;
 
 		try
 		{
-			if (myDB.isClosed())
-			{
-				System.out.print("mist");
-			}
 
 			PreparedStatement statement = myDB.prepareStatement(sql);
 
@@ -105,7 +101,6 @@ public class TestDB extends Database
 		}
 		catch (SQLException e)
 		{
-
 			e.printStackTrace();
 		}
 
@@ -134,7 +129,6 @@ public class TestDB extends Database
 		}
 		catch (SQLException e)
 		{
-
 			e.printStackTrace();
 		}
 
@@ -144,6 +138,7 @@ public class TestDB extends Database
 	@Override
 	public void execSQL(String sql)
 	{
+		if (myDB == null) return;
 		Statement statement;
 		try
 		{
@@ -161,6 +156,7 @@ public class TestDB extends Database
 	@Override
 	public long update(String tablename, Parameters val, String whereClause, String[] whereArgs)
 	{
+		if (myDB == null) return 0;
 
 		StringBuilder sql = new StringBuilder();
 
@@ -219,7 +215,7 @@ public class TestDB extends Database
 	@Override
 	public long insert(String tablename, Parameters val)
 	{
-
+		if (myDB == null) return 0;
 		StringBuilder sql = new StringBuilder();
 
 		sql.append("insert into ");
@@ -263,7 +259,8 @@ public class TestDB extends Database
 				st.setObject(j, entry.getValue());
 			}
 
-			return st.executeUpdate();
+			// return st.executeUpdate();
+			return st.execute() ? 0 : 1;
 
 		}
 		catch (SQLException e)
@@ -276,7 +273,7 @@ public class TestDB extends Database
 	@Override
 	public long delete(String tablename, String whereClause, String[] whereArgs)
 	{
-
+		if (myDB == null) return 0;
 		StringBuilder sql = new StringBuilder();
 
 		sql.append("delete from ");
@@ -315,7 +312,7 @@ public class TestDB extends Database
 	{
 		try
 		{
-			myDB.setAutoCommit(false);
+			if (myDB != null) myDB.setAutoCommit(false);
 		}
 		catch (SQLException e)
 		{
@@ -329,11 +326,10 @@ public class TestDB extends Database
 	{
 		try
 		{
-			myDB.commit();
+			if (myDB != null) myDB.commit();
 		}
 		catch (SQLException e)
 		{
-
 			e.printStackTrace();
 		}
 	}
@@ -343,11 +339,10 @@ public class TestDB extends Database
 	{
 		try
 		{
-			myDB.setAutoCommit(true);
+			if (myDB != null) myDB.setAutoCommit(true);
 		}
 		catch (SQLException e)
 		{
-
 			e.printStackTrace();
 		}
 
@@ -356,6 +351,8 @@ public class TestDB extends Database
 	@Override
 	public long insertWithConflictReplace(String tablename, Parameters val)
 	{
+		if (myDB == null) return 0;
+
 		StringBuilder sql = new StringBuilder();
 
 		sql.append("insert OR REPLACE into ");
@@ -412,6 +409,8 @@ public class TestDB extends Database
 	@Override
 	public long insertWithConflictIgnore(String tablename, Parameters val)
 	{
+		if (myDB == null) return 0;
+
 		StringBuilder sql = new StringBuilder();
 
 		sql.append("insert OR IGNORE into ");
@@ -468,6 +467,9 @@ public class TestDB extends Database
 	@Override
 	public int getCacheCountInDB(String filename)
 	{
+
+		if (myDB == null) return 0;
+
 		int count = 0;
 		Connection myDB = null;
 		try
@@ -483,7 +485,7 @@ public class TestDB extends Database
 		}
 		catch (SQLException e)
 		{
-			e.getMessage();
+			// String s = e.getMessage();
 		}
 		return count;
 	}
